exercises/requirements.txt
langchain-openai>=0.0.5
langchain-core>=0.1.15
langgraph>=0.0.20
jupyter>=1.0.0
python-dotenv>=1.0.0 